rename: 새 구조 설계에 맞춰서 수정
