+++ lib/app_router.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:office_shopping_mall/core/constants/app_constants.dart';
 import 'package:office_shopping_mall/core/constants/app_routes.dart';
-import 'package:office_shopping_mall/feature/auth/ui/login_screen.dart';
-import 'package:office_shopping_mall/feature/auth/ui/signup_screen.dart';
 import 'package:office_shopping_mall/feature/home/ui/home_screen.dart';
 import 'package:office_shopping_mall/feature/mypage/ui/mypage_screen.dart';
 import 'package:office_shopping_mall/feature/notification/ui/notification_screen.dart';
@@ -12,6 +10,8 @@ import 'package:office_shopping_mall/feature/preference/ui/preference_screen.dar
 import 'package:office_shopping_mall/feature/setting/ui/dest_setting_screen.dart';
 import 'package:office_shopping_mall/feature/setting/ui/pw_setting.dart';
 
+import 'feature/auth/presentation/login_screen.dart';
+import 'feature/auth/presentation/signup_screen.dart';
 import 'feature/category/ui/category_screen.dart';
 import 'feature/product/ui/product_detail_screen.dart';
 import 'feature/product/ui/product_list_screen.dart';
